Complete Trixie 64-bit OS migration with libcamera compatibility
## Key Changes

- **libcam.cpp**: Add defensive checks for draft controls (AePrecaptureTrigger,
  NoiseReductionMode, etc.) to gracefully skip unavailable controls on 64-bit OS
- **libcam.cpp**: Accept NV12 pixel format as valid alternative to YUV420 for
  motion detection (both use identical Y-plane)
- **scripts/pi5-build.sh**: Update build configuration for Trixie 64-bit
- **scripts/pi5-setup.sh**: Improve Pi 5 setup script with proper dependencies
- **libcam.hpp**: Minor compatibility adjustments
- **motion.cpp**: Update for 64-bit compatibility
- **video_v4l2.cpp**: Fix compatibility issues with 64-bit kernel
- **CLAUDE.md**: Document Trixie migration completion

## Context

Motion daemon now fully supports Debian Trixie (64-bit) on Raspberry Pi 5 with
Camera Module v3. The migration maintains compatibility with the proven motion
detection algorithm while adapting libcamera integration for the newer OS.

+++ src/libcam.hpp
@@ -33,7 +33,7 @@
         /* Buffers and sizes for planes of image*/
         struct ctx_imgmap {
             uint8_t *buf;
-            int     bufsz;
+            size_t  bufsz;  /* Changed to size_t for 64-bit safety */
         };
 
         /* Request with associated buffer index for multi-buffer support */
